refactor: extracted FallenList component from all-fallen-page into its own file, refactor: added a URL generlizer function
