+++ app/all-fallen/page.jsx
@@ -1,21 +1,21 @@
-import { Suspense, use } from "react";
+import { Suspense } from "react";
+import FallenList from "./FallenList";
 import styles from "./page.module.scss";
 import Button from "../components/Button";
+import { getBaseUrl } from "@/lib/baseUrl";
 import HobbyTag from "../components/HobbyTag";
 import SearchInput from "../components/SearchInput";
-import ProfileCard from "../components/ProfileCard";
 import CustomBubble from "../components/CustomBubble";
 import TitleDivider from "../components/TitleDivider";
 
 export default async function AllFallenPage({ searchParams }) {
   const q = (await searchParams).q || "";
+
+  const baseUrl = getBaseUrl();
+  const fallenPromise = (await fetch(`${baseUrl}/api/fallen?q=${q}`)).json();
+
   // TODO: Replace dummy data in real hobbies from API
   const hobbies = ["טניס", "שירה", "ריצה", "אפיה", "סריגה", "שחיה"];
-
-  // TODO: Generlize base URL depending on the environment
-  const fallenPromise = (
-    await fetch(`http://localhost:3000/api/fallen?q=${q}`)
-  ).json();
 
   return (
     <>
@@ -42,21 +42,6 @@ export default async function AllFallenPage({ searchParams }) {
           <FallenList fallenPromise={fallenPromise} />
         </Suspense>
       </div>
-    </>
-  );
-}
-
-// TODO: Consider moving to a separate file
-function FallenList({ fallenPromise }) {
-  const fallen = use(fallenPromise);
-
-  return (
-    <>
-      {fallen.map((fallen) => (
-        <div className={styles.cardBackground} key={fallen.id}>
-          <ProfileCard fallen={fallen} />
-        </div>
-      ))}
     </>
   );
 }
